refactor: break down WhitelistUserAgentsCard
