+++ app/routes/dashboard.apps.$appId.security/utils/utils.ts
@@ -1,0 +1,6 @@
+export const addIfMissing = (item: string, arr: string[]) => {
+  if (arr.indexOf(item) !== -1) {
+    return arr
+  }
+  return [...arr, item]
+}
